+++ packages/wallets/src/signers/non-custodial/ncs-signer.ts
@@ -9,9 +9,9 @@ import { AuthRejectedError } from "../types";
 import { NcsIframeManager } from "./ncs-iframe-manager";
 import { validateAPIKey } from "@crossmint/common-sdk-base";
 import type { SignerOutputEvent } from "@crossmint/client-signers";
-import { getShadowSigner, hasShadowSigner, ShadowSignerData } from "@/utils/shadow-signer";
-import { Chain } from "@/chains/chains";
-import { ExternalWalletSigner } from "../external-wallet-signer";
+import { getShadowSigner, hasShadowSigner } from "@/signers/shadow-signer";
+import type { Chain } from "@/chains/chains";
+import type { ExternalWalletSigner } from "../external-wallet-signer";
 
 export abstract class NonCustodialSigner implements Signer {
     public readonly type: "email" | "phone";
